+++ DropletsInMotion/Application/Factories/IAgentFactory.cs
@@ -5,6 +5,6 @@ namespace DropletsInMotion.Application.Factories;
 public interface IAgentFactory
 {
     Agent CreateAgent(string dropletName, int positionX, int positionY, double volume);
-    Agent CreateAgent(string dropletName, int positionX, int positionY, double volume, byte substanceId);
-    Agent CreateAgent(string dropletName, int positionX, int positionY, double volume, byte substanceId, LinkedList<(int x, int y)> snakeBody);
+    Agent CreateAgent(string dropletName, int positionX, int positionY, double volume, int substanceId);
+    Agent CreateAgent(string dropletName, int positionX, int positionY, double volume, int substanceId, LinkedList<(int x, int y)> snakeBody);
 }
